pyproject.toml
[project]
name = "view-django"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.10",
    "django-admin-interface>=0.32.0",
    "django-extensions>=4.1",
    "django-silk[formatting]>=5.4.3",
    "djangorestframework>=3.16.1",
]
